Add changes to bootstrap tool for migrating helix resources to Full auto
