+++ ambry-api/src/main/java/com/github/ambry/clustermap/RebalanceMode.java
@@ -1,0 +1,28 @@
+/**
+ * Copyright 2016 LinkedIn Corp. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ */
+package com.github.ambry.clustermap;
+
+/**
+ * Re-balancing mode used for Helix based cluster manager.
+ */
+public enum RebalanceMode {
+  /**
+   * Helix will automatically determine the location and state of each replica
+   */
+  SEMI_AUTO,
+  /**
+   * Helix will take the location of each replica and automatically determine the state
+   */
+  FULL_AUTO
+}
