+++ ambry-clustermap/src/test/java/com/github/ambry/clustermap/MockHelixCluster.java
@@ -66,7 +66,8 @@ public class MockHelixCluster {
     dataCenterToZkAddress = parseDcJsonAndPopulateDcInfo(jsonString);
     HelixBootstrapUpgradeUtil.bootstrapOrUpgrade(hardwareLayoutPath, partitionLayoutPath, zkLayoutPath, clusterName,
         "all", MAX_PARTITIONS_IN_ONE_RESOURCE, false, false, helixAdminFactory, false,
-        ClusterMapConfig.DEFAULT_STATE_MODEL_DEF, BootstrapCluster, DataNodeConfigSourceType.INSTANCE_CONFIG, false);
+        ClusterMapConfig.DEFAULT_STATE_MODEL_DEF, BootstrapCluster, DataNodeConfigSourceType.INSTANCE_CONFIG, false,
+        RebalanceMode.SEMI_AUTO);
     this.clusterName = clusterName;
     this.localHelixAdmin =
         helixAdminFactory.getHelixAdmin(dataCenterToZkAddress.get(localDC).getZkConnectStrs().get(0));
@@ -80,7 +81,8 @@ public class MockHelixCluster {
   void upgradeWithNewHardwareLayout(String hardwareLayoutPath) throws Exception {
     HelixBootstrapUpgradeUtil.bootstrapOrUpgrade(hardwareLayoutPath, partitionLayoutPath, zkLayoutPath, clusterName,
         "all", MAX_PARTITIONS_IN_ONE_RESOURCE, false, false, helixAdminFactory, false,
-        ClusterMapConfig.DEFAULT_STATE_MODEL_DEF, BootstrapCluster, DataNodeConfigSourceType.INSTANCE_CONFIG, false);
+        ClusterMapConfig.DEFAULT_STATE_MODEL_DEF, BootstrapCluster, DataNodeConfigSourceType.INSTANCE_CONFIG, false,
+        RebalanceMode.SEMI_AUTO);
     triggerInstanceConfigChangeNotification();
   }
 
@@ -94,7 +96,7 @@ public class MockHelixCluster {
       HelixBootstrapUpgradeUtil.HelixAdminOperation adminOperation) throws Exception {
     HelixBootstrapUpgradeUtil.bootstrapOrUpgrade(hardwareLayoutPath, partitionLayoutPath, zkLayoutPath, clusterName,
         "all", 3, false, false, helixAdminFactory, false, ClusterMapConfig.DEFAULT_STATE_MODEL_DEF, adminOperation,
-        DataNodeConfigSourceType.INSTANCE_CONFIG, false);
+        DataNodeConfigSourceType.INSTANCE_CONFIG, false, RebalanceMode.SEMI_AUTO);
     triggerInstanceConfigChangeNotification();
   }
